Modify the zero_current component
Now only sets the velocity of a species based on the net current
from other species (similar to quasineutral for the density).

Calculation of the electric field has to move to a separate
component.

This is needed because electron velocity is needed to calculate
collisions, and forces on electrons due to collisions are needed
to calculate the electric field.