Fixed underlying bug when flooring
If density was floored it would blow up at the boundary. The same
thing happened when the density was copied (without flooring).
Caused because the density was not being updated in
evolve_density.finally(). Without copying the memory was the same
as the field stored, but when copied the memory was different.