Fix bug in electron collisions
Introduced in 562f191d1 (July 26th 2022), an `if` rather than `else if`
meant that additional collisions were included between ions and electrons if
`electron_neutral` was `true` (the default).

This increased the electron collision rate, decreasing electron parallel
heat conduction, and increasing the electron temperature.

This adds a new integrated test, `1D-recycling`, that is the same as
the example of that name but with fewer cells (50). That is sufficient
to check the approximate ion and electron temperatures, and would have
caught the bug. The new test should catch changes to electron-ion collisions
in future.