vorticity: Fix diamagnetic_polarisation energy transfer
Should include both parallel and diamagnetic current, but only
diamagnetic term was included in the transfer of energy to/from
ion pressure.

Unfortunately calculating this term depends on current, that
depends on sheath boundaries, that depends on potential. It
therefore can't be calculated in `vorticity` unless assumptions
are made on parallel current at the boundary (e.g. zero-gradient).

Instead, these terms are now calculated in the `polarisation_drift`
component, because they are connected to the divergence of the
polarisation drift velocity.

Fix only affects simulations with vorticity equation and hot ions.