Div_par_fvv: Change how boundary dissipation is done
Previously the damping depended on both the difference between
reconstructed and cell-center velocity AND differences in density.
In cases of large (supersonic) flow, depending on the density profile
this numerical damping term could result in flows accelerating.
By removing the dependence on the density, and just damping the velocity,
this seems to avoid some sheath boundary-induced numerical instabilities.