neutral_parallel_momentum: Add toroidal_slip option
Work in progress. Runs but not properly tested.

Sets the parallel flow at the wall so that the combined flow into the
target is zero, using a constant neutral pressure gradient.  This
allows a net flow along the target.

An alternative would be to set e.g. a Neumann condition on the
parallel flow, and use that to set the neutral pressure gradient
at the wall.